refactor primary id length checking

+++ rust/otap-dataflow/crates/pdata/src/error.rs
@@ -205,7 +205,7 @@ pub enum Error {
     TooManyItems {
         payload_type: ArrowPayloadType,
         count: usize,
-        max: usize,
+        max: u64,
         message: String,
     },
